[fact] Refactor project layout
