+++ src/main/java/se/kth/spork/cli/Cli.java
@@ -8,6 +8,7 @@ import picocli.CommandLine;
 import se.kth.spork.spoon.Compare;
 import se.kth.spork.spoon.Parser;
 import se.kth.spork.spoon.Spoon3dmMerge;
+import se.kth.spork.spoon.printer.PrinterPreprocessor;
 import se.kth.spork.util.LineBasedMerge;
 import se.kth.spork.util.Pair;
 import spoon.reflect.declaration.*;
