+++ src/main/java/se/kth/spork/spoon/StructuralConflict.java
@@ -1,10 +1,10 @@
 package se.kth.spork.spoon;
 
 import se.kth.spork.base3dm.Pcs;
+import se.kth.spork.spoon.wrappers.SpoonNode;
 import spoon.reflect.declaration.CtElement;
 
-import java.util.List;
-import java.util.Objects;
+import java.util.*;
 
 /**
  * A simple class that provides some information on a structural conflict. Meant to be put as metadata on a conflict
@@ -42,5 +42,25 @@ public class StructuralConflict {
         return !Objects.equals(left.getSuccessor(), right.getSuccessor()) &&
                 Objects.equals(left.getPredecessor(), right.getPredecessor()) &&
                 Objects.equals(left.getRoot(), right.getRoot());
+    }
+
+    public static Set<SpoonNode>
+    extractRootConflictingNodes(Map<Pcs<SpoonNode>, Set<Pcs<SpoonNode>>> structuralConflicts) {
+        Set<SpoonNode> toIgnore = new HashSet<>();
+
+        for (Map.Entry<Pcs<SpoonNode>, Set<Pcs<SpoonNode>>> entry : structuralConflicts.entrySet()) {
+            Pcs<SpoonNode> pcs = entry.getKey();
+            for (Pcs<SpoonNode> other : entry.getValue()) {
+                if (isRootConflict(pcs, other)) {
+                    if (pcs.getPredecessor().equals(other.getPredecessor())) {
+                        toIgnore.add(other.getPredecessor());
+                    }
+                    if (pcs.getSuccessor().equals(other.getSuccessor())) {
+                        toIgnore.add(other.getSuccessor());
+                    }
+                }
+            }
+        }
+        return toIgnore;
     }
 }
