+++ src/main/java/se/kth/spork/spoon/pcsinterpreter/ContentMerger.java
@@ -1,6 +1,10 @@
-package se.kth.spork.spoon;
+package se.kth.spork.spoon.pcsinterpreter;
 
 import se.kth.spork.base3dm.Content;
+import se.kth.spork.spoon.ContentConflict;
+import se.kth.spork.spoon.wrappers.RoledValue;
+import se.kth.spork.spoon.wrappers.RoledValues;
+import se.kth.spork.spoon.wrappers.SpoonNode;
 import se.kth.spork.util.LineBasedMerge;
 import se.kth.spork.util.Pair;
 import se.kth.spork.util.Triple;
