+++ src/main/java/se/kth/spork/spoon/pcsinterpreter/PcsInterpreter.java
@@ -1,6 +1,10 @@
-package se.kth.spork.spoon;
+package se.kth.spork.spoon.pcsinterpreter;
 
 import se.kth.spork.base3dm.ChangeSet;
+import se.kth.spork.spoon.wrappers.NodeFactory;
+import se.kth.spork.spoon.wrappers.RoledValues;
+import se.kth.spork.spoon.matching.SpoonMapping;
+import se.kth.spork.spoon.wrappers.SpoonNode;
 import se.kth.spork.util.Pair;
 import spoon.reflect.declaration.CtElement;
