+++ src/main/java/se/kth/spork/spoon/pcsinterpreter/SporkTree.java
@@ -1,11 +1,14 @@
-package se.kth.spork.spoon;
+package se.kth.spork.spoon.pcsinterpreter;
 
 import se.kth.spork.base3dm.Content;
 import se.kth.spork.base3dm.Pcs;
 import se.kth.spork.base3dm.Revision;
+import se.kth.spork.spoon.wrappers.NodeFactory;
+import se.kth.spork.spoon.wrappers.RoledValues;
+import se.kth.spork.spoon.wrappers.SpoonNode;
+import se.kth.spork.spoon.StructuralConflict;
 
 import java.util.*;
-import java.util.stream.Collectors;
 
 
 /**
