+++ src/main/java/se/kth/spork/spoon/printer/OperatorHelper.java
@@ -19,7 +19,7 @@
  * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
  * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package se.kth.spork.cli;
+package se.kth.spork.spoon.printer;
 
 import spoon.SpoonException;
 import spoon.reflect.code.BinaryOperatorKind;
