+++ src/main/java/se/kth/spork/spoon/wrappers/SpoonNode.java
@@ -1,4 +1,4 @@
-package se.kth.spork.spoon;
+package se.kth.spork.spoon.wrappers;
 
 import se.kth.spork.base3dm.ListNode;
 import se.kth.spork.base3dm.Revision;
