+++ src/test/java/se/kth/spork/Util.java
@@ -5,19 +5,12 @@ import gumtree.spoon.builder.SpoonGumTreeBuilder;
 import org.junit.jupiter.api.extension.ExtensionContext;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.ArgumentsProvider;
-import se.kth.spork.cli.SporkPrettyPrinter;
+import se.kth.spork.spoon.printer.SporkPrettyPrinter;
 import spoon.Launcher;
-import spoon.compiler.Environment;
-import spoon.reflect.CtModel;
 import spoon.reflect.declaration.CtClass;
-import spoon.reflect.declaration.CtCompilationUnit;
-import spoon.reflect.declaration.CtModule;
-import spoon.support.compiler.FileSystemFile;
-import spoon.support.sniper.SniperJavaPrettyPrinter;
 
 import java.io.File;
 import java.io.IOException;
-import java.lang.reflect.Array;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
